+++ changelog.md
@@ -1,4 +1,6 @@
 # Icon Ninja Changelog
+## v1.0.2
+- Fixed issues and incompatibilities relating to touch
 ## v1.0.1
 - Added margin of +/-5px on icons
 - Changed intersection algorithm